1.2

+++ src/Traits/FileUploadable.php
@@ -1,11 +1,11 @@
 <?php
 /**
- * CodersStudio 2019
- * https://coders.studio
- * info@coders.studio
+ * Eugeny Nosenko 2021
+ * https://toprogram.ru
+ * info@toprogram.ru
  */
 
-namespace CodersStudio\CRUD\Traits;
+namespace Nos\CRUD\Traits;
 
 use Illuminate\Http\Request;
 use Illuminate\Database\Eloquent\Model;
@@ -20,13 +20,15 @@ trait FileUploadable
 	 */
     protected function upload(Request $request, Model $model, string $collectionName = 'media_collection'):void
     {
-	    $mediaCollection = $request->get($collectionName, ['name' => '', 'files' => []]);
-	    
-    	foreach($mediaCollection['files'] as $key => $file) {
-    		$model->addMedia(storage_path('app/' . $file['path']))->toMediaCollection($mediaCollection['name']);
-    	}
-	    foreach($mediaCollection['removedFiles'] as $key => $file) {
-    		$model->deleteMedia($file['id']);
-    	}
+        if($request->exists($collectionName)) {
+            $mediaCollection = $request->get($collectionName, ['name' => '', 'files' => []]);
+
+            foreach ($mediaCollection['files'] as $key => $file) {
+                $model->addMedia(storage_path('app/' . $file['path']))->toMediaCollection($mediaCollection['name']);
+            }
+            foreach ($mediaCollection['removedFiles'] as $key => $file) {
+                $model->deleteMedia($file['id']);
+            }
+        }
     }
 }
